+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Bug #3: Change visibility poperty `attributes` to `public` in tests (@terabytesoftw)
 - Enh #4: Add `HasAriaCurrent` trait and `ariaCurrent()` method (@terabytesoftw)
+- Bug #5: Remove dead code in `HasAriaCurrent` trait (@terabytesoftw)
 
 ## 0.1.1 March 7, 2024